% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sign_traits.R
\name{get_sign_traits}
\alias{get_sign_traits}
\title{Tells a daily horoscope}
\usage{
get_sign_traits(zodiac)
}
\arguments{
\item{zodiac}{the person's zodiac sign}
}
\value{
A string with all sign information
}
\description{
Tells a daily horoscope
}
