% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_zodiac_sign.R
\docType{data}
\name{url}
\alias{url}
\title{Gives zodiac sign}
\format{
An object of class \code{character} of length 1.
}
\usage{
url
}
\arguments{
\item{user_date}{the person's birthday}
}
\value{
the zodiac sign
}
\description{
Gives zodiac sign
}
\keyword{datasets}
